requirements.txt
absl-py==2.1.0
appnope==0.1.4
asttokens==3.0.0
attrs==24.3.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
build==1.2.2.post1
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
cloudpickle==3.1.0
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.10.0
grpcio==1.68.1
gym==0.26.2
gym-notices==0.0.8
-e git+https://github.com/dellalibera/gym-backgammon.git@cdeee45b73779809f938ae95a80d409fd6bfba7e#egg=gym_backgammon
idna==3.10
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.8
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mistune==3.1.0
mpmath==1.3.0
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.0
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
pip-tools==7.4.1
pipreqs==0.5.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==5.29.1
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyglet==2.0.20
Pygments==2.18.0
pyparsing==3.2.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
six==1.17.0
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.1
tb-nightly==2.19.0a20241218
tensorboard-data-server==0.7.2
tinycss2==1.4.0
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
yarg==0.1.9
